1.完善WS推HTTP拉的关系;2.调整聊天Seq的维护与处理; 3.调整用户与群组的主键为雪花ID,替换原默认主键; 4.完善conversation_id的建立与使用在消息流中

+++ DB/online.go
@@ -7,19 +7,19 @@ import (
 )
 
 // SetOnline 缓存在线状态
-func SetOnline(userID uint) {
+func SetOnline(userID int64) {
 	var online = 1
 
 	RDB.HSetNX(context.TODO(), "users:online", strconv.Itoa(int(userID)), online)
 }
 
 // SetOffline 删除缓存的在线状态
-func SetOffline(userID uint) {
+func SetOffline(userID int64) {
 	RDB.HDel(context.TODO(), "users:online", strconv.Itoa(int(userID)))
 }
 
 // CheckOnline 检查是否在线
-func CheckOnline(userID uint) (online bool) {
+func CheckOnline(userID int64) (online bool) {
 	res := RDB.HGet(context.TODO(), "users:online", strconv.Itoa(int(userID)))
 	_, err := res.Result()
 	if err != nil {
@@ -29,6 +29,6 @@ func CheckOnline(userID uint) (online bool) {
 }
 
 // SetLastOnline 更新上次在线时间
-func SetLastOnline(userID uint) {
+func SetLastOnline(userID int64) {
 	MySQL.Table("users").Where("id=?", userID).Update("last_online", time.Now())
 }
