+++ DB/user-info.go
@@ -1,7 +1,7 @@
 package DB
 
 // InsertUserInfo 初始化用户的info
-func InsertUserInfo(userID uint) error {
+func InsertUserInfo(userID int64) error {
 	var userInfo = UserInfo{
 		UserID: userID,
 	}
@@ -10,7 +10,7 @@ func InsertUserInfo(userID uint) error {
 }
 
 // UpdateUserInfo 更新用户的Info
-func UpdateUserInfo(userID uint, postform UserInfo) error {
+func UpdateUserInfo(userID int64, postform UserInfo) error {
 	if err := MySQL.Table("user_infos").Where("user_id=?", userID).Updates(&postform).Error; err != nil {
 		return err
 	}
@@ -18,7 +18,7 @@ func UpdateUserInfo(userID uint, postform UserInfo) error {
 }
 
 // UpdateUserName 更新用户的昵称
-func UpdateUserName(userID uint, New string) error {
+func UpdateUserName(userID int64, New string) error {
 	if err := MySQL.Table("users").Where("id=?", userID).Update("nickname", New).Error; err != nil {
 		return err
 	}
